refactor(download): remove the pridepy/files/ back-compat shim
The pridepy/files/files.py shim (Files = Client re-export) was load-bearing
only for the test suite — the CLI already imports from pridepy.download.client.
Updated the 9 test files to 'from pridepy.download.client import Client as Files'
(keeps the local Files alias so test bodies are otherwise untouched) and
deleted the pridepy/files/ package entirely.

Breaking change for any downstream code doing 'from pridepy.files.files
import Files' — use 'from pridepy.download.client import Client' instead.
Pre-1.0 cleanup; the class was already renamed Files -> Client in this PR.

Also updated two stale docstrings (transport.py, client.py) that referenced
the old pridepy.files.files path.

68 passed, 4 skipped. flake8 (E9,F63,F7,F82) clean. Smoke MD5 unchanged.
