+++ pridepy/tests/test_jpost_files.py
@@ -3,7 +3,7 @@ import tempfile
 from unittest import TestCase
 from unittest.mock import MagicMock, patch
 
-from pridepy.files.files import Files
+from pridepy.download.client import Client as Files
 from pridepy.download import transport
 from pridepy.download.jpost import JpostProvider
