+++ pridepy/tests/test_massive_files.py
@@ -2,7 +2,7 @@ import tempfile
 from unittest import TestCase
 from unittest.mock import patch
 
-from pridepy.files.files import Files
+from pridepy.download.client import Client as Files
 from pridepy.download import transport
 from pridepy.download.massive import MassiveProvider
