+++ pridepy/tests/test_search.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from pridepy.files.files import Files
+from pridepy.download.client import Client as Files
 from pridepy.project.project import Project
 from pridepy.util.api_handling import Util
 import logging
